wip(webapp): restructuration
