+++ src/app/core/adapter/authentication-firebase.service.ts
