--- src/app/core/repository/user.service.spec.ts
@@ -1,19 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { UserService } from './user.service';
-import { provideHttpClient } from '@angular/common/http';
-
-describe('UserService', () => {
-  let service: UserService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [ provideHttpClient()]
-    });
-    service = TestBed.inject(UserService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
