+++ src/app/membership/dashboard/dashboard.page.component.html
@@ -1,0 +1,3 @@
+<h1 class="text-center mt-5">Hello, {{ store.user()?.name }} ! 👋</h1>
+
+<p class="text-center">Email : {{ store.user()?.email }}</p>
