+++ src/app/visitor/signup/domain/register-user.use-case.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { RegisterUserUseCaseService } from './register-user.use-case.service';
-import { AuthenticationService } from '@app/core/authentication.service';
-import { UserService } from '@app/core/repository/user.service';
+import { AuthenticationService } from '@app/core/port/authentication.service';
+import { UserService } from '@app/core/port/user.service';
 import { UserStore } from '@app/core/store/user.store';
 import { Router } from '@angular/router';
